Fixed C++ linking support
Without __BEGIN_DECLS and __END_DECLS, the C++ compiler will mangle the function names and fail to link. With those added, when using a C++ compiler, it will automatically wrap the declarations in extern "C" { } which prevents the mangling, but will not insert anything when using a C compiler (this is used in all of the other kos-ports sound libraries to allow C++ support).